rubocop指摘の修正
