+++ 07.ls_object/class_relation
@@ -1,0 +1,23 @@
+classDiagram
+    class Segment{
+      +Option option
+      +Array~File files
+      +output()
+    }
+    class DetailSegment{
+      +Array shots
+      +output() int
+      +strike?() bool
+      +spare?() bool
+    }
+    class Game{
+      +Array marks
+      +Array frames
+      +split_into_frames() Array~Frame~
+      +score() int
+      +strike_bonus(int) int
+      +spare_bonus(int) int
+    }
+
+    Game "1" --> "1..10" Frame : has
+    Frame "1" --> "1..3" Shot : contains
